requirements.txt
# =======================================
# FLASK MACHINE LEARNING PLATFORM
# Dependencias del Proyecto
# =======================================

# ---- CORE FLASK ----
Flask>=2.3.3,<3.0.0
Flask-WTF>=1.1.1,<2.0.0
python-dotenv>=1.0.0,<2.0.0
WTForms>=3.0.1,<4.0.0
Werkzeug>=2.3.7,<3.0.0

# ---- MACHINE LEARNING ----
scikit-learn>=1.3.0,<2.0.0
joblib>=1.3.0,<2.0.0
mlxtend>=0.23.0,<0.24.0

# ---- DATA PROCESSING ----
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<3.0.0
openpyxl>=3.1.0,<4.0.0

# ---- VISUALIZATION ----
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<1.0.0

# ---- UTILIDADES ----
psutil>=5.9.0,<8.0.0
python-dateutil>=2.8.0,<3.0.0
charset-normalizer>=3.0.0,<4.0.0
chardet>=5.0.0,<6.0.0
# Nota: uuid es parte de la librería estándar de Python, no requiere instalación

# ---- DESARROLLO (OPCIONAL) ----
# pytest>=7.4.0,<8.0.0          # Testing
# black>=23.0.0,<24.0.0          # Formateo de código
# flake8>=6.0.0,<7.0.0           # Linting